requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
emoji==2.15.0
filetype==1.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
langdetect==1.0.9
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.3.4
olefile==0.47
packaging==25.0
psutil==7.1.3
pypandoc==1.13
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pypdf==6.2.0
python-docx==1.2.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.18
unstructured-client==0.42.3
urllib3==2.5.0
webencodings==0.5.1
wrapt==2.0.1
